requirements.txt
requests>=2.31.0,<3.0
playwright>=1.36.0,<2.0
TikTokApi
pandas
pyyaml